requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Frontend
streamlit>=1.28.0
altair>=5.0.0
pandas>=2.1.0

# HTTP Client
requests>=2.31.0

# Database
sqlalchemy>=2.0.0

# LangChain & LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# Langfuse (Observability)
langfuse>=2.0.0

# Configuration & Settings
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Vector Store
faiss-cpu>=1.7.4

# Document Processing
python-docx>=0.8.11
pypdf>=4.0.0

